NTFS: First level of support for large file records - copes with files that have many attributes, but does not yet split large attributes.

+++ src/Ntfs/NtfsFileSystem.cs
@@ -432,7 +432,11 @@ namespace DiscUtils.Ntfs
                 File file = GetFile(dirEntry.Reference);
 
                 parentDir.RemoveEntry(dirEntry);
-                file.Delete();
+
+                if (file.HardLinkCount == 0)
+                {
+                    file.Delete();
+                }
             }
         }
